Light the party, and load the box
The two sides of a negotiation are lit by the same idea from opposite ends. The
enemies go black except the ones the offer was put to; the party goes dim except
whoever is speaking, who comes up to full and gets there four times as fast.
The speaker is then put on one of two scripts out of its model's table,
according to a byte in its own record.

The box's graphics arrive the other way round from everything else: unpacked to
scratch in one piece, then sent to VRAM as two - the first 0x200 bytes are a
256-entry palette and the rest the frame's tiles, which share the message
windows' page column and move with it. Bit 0x20 of the flags is what stops it
happening twice.

Also the voice request the negotiation puts in - what sort of line, which line,
and which enemies it is meant for, the last taken from the offer so a line
reaches everyone taking part rather than one demon.

Two struct fields settle with them: an object's +4 is the model's table of
scripts, and an actor's +0xC4 is which of two that table supplies. The second
was almost recorded at +0x88 - g_btl_actors starts at 0x8005D654, not at the
0x8005D690 the calls happen to be written against, and the two are 0x3C apart.

+++ src/p1-jp/btlp/voice.c
@@ -23,7 +23,7 @@ extern short g_btl_offer_slot;
 
 void BtlQueueVoice(u_short line, int alt)
 {
-    short kind;
+    int kind;
 
     /* Written once outright and once again from the branch. The first store
        looks redundant and is in the original. */
